Refactor folders navigation
Folders navigation are now URL based. Changes in the URL (the last
fragment is the folder id) are now watched and the view is updated
accordingly.
Also added a new field `paths` to folder json corresponding to all
ancestors of the folder.
Automatic URL path correction if any folder before the target folder was
changed (moved for example).
Support opening a folder and a document simultaneously (automatic URL
correction is disabled in that case).
New separated routing for /home and /home/* as a workaround.