Fix users able to change to an old used address.
Fix org not checked against old org (check on slug).

Added validators to fields `slug` (FTLOrg) and `email` (FTLUser). Be aware that validators are not called automatically.

Removed `clean_email` from FTLUserCreationForm because the validator is automatically used.

Simplified code in command `disable_account`. Accounts are either deleted completely from DB (and can be recreated with same org or email) or accounts are disabled (and can't be recreated).